+++ breeze-core/src/main/kotlin/annotations/MarkAnnotations.kt
@@ -1,7 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
- * Breeze is blowing...
- ******************************************************************************/
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
 
 package com.windea.breezeframework.core.annotations
 
@@ -9,8 +7,9 @@ import kotlin.annotation.AnnotationRetention.*
 import kotlin.annotation.AnnotationTarget.*
 
 /**
+ * 未实现的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Items that is not implemented (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable
@@ -21,8 +20,9 @@ annotation class NotImplemented(
 )
 
 /**
+ * 未包含的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Items that is not included (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable
@@ -33,8 +33,9 @@ annotation class NotIncluded(
 )
 
 /**
+ * 未测试的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Item that is not tested (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable
@@ -45,8 +46,9 @@ annotation class NotTested(
 )
 
 /**
+ * 未优化的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Item that is not optimized (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable
@@ -57,8 +59,9 @@ annotation class NotOptimized(
 )
 
 /**
+ * 不推荐的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Item that is not recommended (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable
@@ -69,8 +72,9 @@ annotation class NotRecommended(
 )
 
 /**
- * Item that is not included (yet, always or at most times).
- * @property message additional note information.
+ * 不适用的项（至今为止，总是如此或者在多数情况下）。
+ *
+ * Item that is not suitable (yet, always or at most times).
  */
 @MustBeDocumented
 @Repeatable
@@ -81,8 +85,9 @@ annotation class NotSuitable(
 )
 
 /**
+ * 不可用的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Item that is not usable (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable
@@ -93,8 +98,9 @@ annotation class NotUsable(
 )
 
 /**
+ * 不确定的项（至今为止，总是如此或者在多数情况下）。
+ *
  * Item that is not sure (yet, always or at most times).
- * @property message additional note information.
  */
 @MustBeDocumented
 @Repeatable