+++ breeze-core/src/main/kotlin/annotations/MessageAnnotations.kt
@@ -1,15 +1,14 @@
-/*******************************************************************************
- * Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
- * Breeze is blowing...
- ******************************************************************************/
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
 
 package com.windea.breezeframework.core.annotations
 
 import kotlin.annotation.AnnotationTarget.*
 
 /**
+ * 为被注解的声明提供一组名字。
+ *
  * Annotation that provide a list of names the to annotated declaration.
- * @property value the name list.
  * */
 @MustBeDocumented
 @Target(CLASS, PROPERTY, FIELD)
@@ -18,9 +17,10 @@ annotation class Name(
 )
 
 /**
+ * 为被注解的声明提供一组别名。
+ *
  * Annotation that provide a list of aliases the to annotated declaration.
- * @property value the alias list.
- * */
+ */
 @MustBeDocumented
 @Target(CLASS, PROPERTY, FIELD)
 annotation class Alias(
@@ -28,9 +28,10 @@ annotation class Alias(
 )
 
 /**
+ * 为被注解的声明提供一组引用。
+ *
  * Annotation that provide a list of references the to annotated declaration.
- * @property value the reference list.
- * */
+ */
 @MustBeDocumented
 @Target(CLASS, PROPERTY, FIELD, FUNCTION, FILE)
 annotation class Reference(