+++ breeze-core/src/main/kotlin/domain/Converter.kt
@@ -1,7 +1,5 @@
-/*******************************************************************************
- * Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
- * Breeze is blowing...
- ******************************************************************************/
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
 
 package com.windea.breezeframework.core.domain
 
@@ -10,3 +8,4 @@ interface Converter<in T: Any,out R: Any>{
 
 	fun convertOrNull(value:T):R? = convert(value)
 }
+