+++ breeze-core/src/main/kotlin/domain/collections/MapEntry.kt
@@ -1,11 +1,13 @@
-/*******************************************************************************
- * Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
- * Breeze is blowing...
- ******************************************************************************/
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
 
 package com.windea.breezeframework.core.domain.collections
 
-/**映射入口。*/
+/**
+ * 映射入口。
+ *
+ * Map entry.
+ */
 data class MapEntry<K, V>(
 	override val key: K,
 	override val value: V