+++ CHANGELOG.md
@@ -406,6 +406,9 @@
 * [X] 实现`<Collection>.deepGetOrElse`
 * [X] 修复`<Collection>.deepFlatten`中的bug
 * [X] 重构和添加`List.collapse`和`Any.expand`
+* [X] 重命名`String.splitMatched`和`String.splitToStrings`
+* [X] 使用java编写常量类。
+* [X] `Color`、`ColorService`、`String.toColor`、`String.toColorOrNull`以及注册默认的css颜色
 
 # 长期